requirements.txt
Flask>=2.2.2
python-dotenv>=0.21.0
gunicorn>=21.0.0
mysql-connector-python>=8.0.33
colorama==0.4.6